Create SES class and module based on Slack discussion
* Create generic SES class with basic send_email function for credit 
card slips use case
* Modify and shift corresponding test
